// apps/web/src/lib/events.generated.ts
export type EventImage = {
  alt: string;
  width: number;
  height: number;
  lqip: string;
  avif: { w: number; url: string }[];
  webp: { w: number; url: string }[];
};